// backend/test/generated/client/types.gen.ts
export type AuthApiKey = {
  id: string;
  userId: string;
  apiKey: string | null;
  isActive: boolean | null;
  createdAt: string;
  updatedAt: string;
  AuthUser?: AuthUser;
};

export type AuthSession = {
  id: string;
  userId: string;
  isActive: boolean | null;
  createdAt: string;
  updatedAt: string;
  AuthUser?: AuthUser;
};

export type ChatDocumentEmbedding = {
  id: string;
  content: string;
  graphContent: string | null;
  metadata: {
    [key: string]: unknown;
  } | null;
  contentHash: string;
  provider: string | null;
  model: string | null;
  createdAt: string;
  updatedAt: string;
  ChatMessageDocumentEmbedding?: Array<ChatMessageDocumentEmbedding>;
};

export type ChatMessageDocumentEmbedding = {
  id: string;
  messageId: string | null;
  embeddingDocumentId: string | null;
  isFound: boolean;
  relevanceScore: number | null;
  createdAt: string;
  updatedAt: string;
  chatHistory?: ChatMessage | null;
  embeddingDocument?: ChatDocumentEmbedding | null;
};

export type ChatLlmRequest = {
  id: string;
  request: string;
  response: string;
  requestLength: number;
  responseLength: number;
  executionTimeMs: number;
  provider: string;
  model: string;
  temperature: number | null;
  isSuccess: boolean;
  errorMessage: string | null;
  dialogId: string | null;
  messageId: string | null;
  createdAt: string;
  updatedAt: string;
  dialog?: ChatDialog | null;
  history?: ChatMessage | null;
};

export type ChatMessage = {
  id: string;
  userId: string;
  question: string;
  answer: string;
  dialogId: string | null;
  isFound: boolean;
  category: string | null;
  transformedQuestion: string | null;
  transformedEmbeddingQuery: string | null;
  provider: string | null;
  model: string | null;
  temperature: number | null;
  isGoodResponse: boolean;
  isBadResponse: boolean;
  trace: {
    [key: string]: unknown;
  } | null;
  constants: {
    [key: string]: unknown;
  } | null;
  isProcessing: boolean;
  questionReceivedAt: string | null;
  answerSentAt: string | null;
  createdAt: string;
  updatedAt: string;
  deletedAt: string | null;
  AuthUser?: AuthUser;
  dialog?: ChatDialog | null;
  ChatMessageDocumentEmbedding?: Array<ChatMessageDocumentEmbedding>;
  ChatLlmRequest?: Array<ChatLlmRequest>;
};

export type ChatDialog = {
  id: string;
  userId: string;
  title: string | null;
  summary: string | null;
  consecutiveFailures: number;
  isFailed: boolean;
  createdAt: string;
  updatedAt: string;
  AuthUser?: AuthUser;
  ChatMessage?: Array<ChatMessage>;
  ChatLlmRequest?: Array<ChatLlmRequest>;
};

export type AuthUser = {
  id: string;
  anonymousId: string | null;
  supabaseUserId: string | null;
  supabaseUserData: {
    [key: string]: unknown;
  } | null;
  isActive: boolean | null;
  createdAt: string;
  updatedAt: string;
  AuthApiKey?: Array<AuthApiKey>;
  AuthSession?: Array<AuthSession>;
  ChatDialog?: Array<ChatDialog>;
  ChatMessage?: Array<ChatMessage>;
};